refactor: move selfdesctruct function and settings to corresponding places
