+++ commands/__init__.py
@@ -1,0 +1,23 @@
+from .alarm_command import alarm
+from .clear_command import clear
+from .diag_command import diag
+from .door_command import door
+from .file_command import file
+from .help_command import help
+from .item_command import item
+from .ping_command import ping
+from .selfdestruct_command import selfdestruct
+from .service_command import service
+
+__all__ = [
+    alarm,
+    clear,
+    diag,
+    door,
+    file,
+    help,
+    item,
+    ping,
+    selfdestruct,
+    service,
+]
