+++ settings.py
@@ -1,0 +1,27 @@
+# Things to set up audio for terminal
+ENABLE_MUSIC = True
+ENABLE_TYPEWRITER = True
+
+# Parameters for slow typer
+SLOW_LINE_SLEEP = 0.15
+SLOW_TYPE_SPEED = 1000
+
+# Self Destruct settings
+SELFDRSTRUCT_TIMER = 600
+SELFDRSTRUCT_PASSWORD_START = "hackme12"
+SELFDRSTRUCT_PASSWORD_STOP = "hackme14"
+SELFDRSTRUCT_ACTIVE = False
+SELFDRSTRUCT_DEAD = False
+
+# sleep time before startup starts working
+START_SLEEP_TIME = 5
+
+# file password settings
+DECRYPTION_IS_SLOW = True
+DECRYPTION_SPEED = 1000
+DECRYPTION_WAV_DIVIDER = 10000
+DECRYPTION_SPEED_WAV = 1000
+DECRYPTION_LAG = 10
+
+# break line for terminal AND (!) diag file
+BREAKLINE = "- - - - - - - - - - - - - - - - - - - - - - -\n"
